+++ CHANGES.rst
@@ -2,7 +2,7 @@
 Revision 0.3.1, released XX-07-2017
 -----------------------------------
 
-- Fixed UTCTimeEncoder CER encoder time string format verification
+- Fixed GeneralizedTime/UTCTime CER/DER codecs to actually get invoked
 
 Revision 0.3.1, released 26-07-2017
 -----------------------------------